Refactor package as hyped add-on
