+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "hyped_serve"
+name = "hyped-serve"
 dynamic = ["version"]
 description = "TODO"
 readme = "README.md"
@@ -23,17 +23,13 @@ dependencies = [
 linting = ["pre-commit"]
 tests = ["pytest", "pytest-cov", "httpx"]
 docs = ["sphinx", "sphinx_rtd_theme"]
-peft = [
-    "peft>=0.9.0",
-    "bitsandbytes>=0.42.0"
-]
 
 [build-system]
 requires = ["setuptools >= 61.0.0", "setuptools_scm>=8.0.0"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-version_file = "src/hyped_serve/__version__.py"
+version_file = "src/hyped/serve/__version__.py"
 
 [tool.black]
 target-version = ["py311"]
