--- src/hyped_serve/__init__.py
@@ -1,2 +1,0 @@
-"""hyped.serve."""
-from hyped.__version__ import __version__, __version_tuple__
